kernel: allow use of fpu in note handlers
Instead of making it illegal to use FPU in note-handler,
we save, disable and mark the fpstate with FPnotify.
Once in emu-trap, we call notefpsave() which will
make a copy of the original register state once
the note-handler returns (which is also exposed as
/proc/n/notefpregs).

This patch applies to arm64, amd64, 386 and most
arm32 kernels. The remaining kernels are left as
an exercise to the reader.