kernel: fix noted() checks for all archs
Theres a potential race allowing to set arbitrary regs/flags,
when the user-space Ureg struct is in a shared segment.
Then another process could modify the flags after our
check passed.

The right way todo it is to use setregisters(), which
copies first and then adjusts flags. And all later checks
use the kernel-copy of the Ureg, ignoring the user-space
one.