kernel: restore kernel fpu state on fault*() error
Theres one special case when handling a page-fault
from syscall handler which accesses user-memory and
we are unable to handle the fault.

In that case, faultamd64()/faultarm64() gotolabel() to
the last waserror() handler, skipping the return back to
trap(), which would have restored the kernel-fpu context.

As we now track the nested FPsave context using FPalloc.link
pointers, we can avoid returning FPalloc* from fpukenter()
and passing the argument to fpukexit().