kernel: explicitely pass MiiPhy* to ethermii operations
The Mii.curphy pointer is just a convenient shorthand
that points to the first enumerated phy found by mii().

If multiple phy's are in use on a single mdio bus,
we want the interface to not depend on Mii.curphy.

This also simplifies the code for most operations,
as curphy checking can be avoided.

All clause22 accesses now pass thru miimir()/miimiw()
(except when enumerating). Which allows us to put some
serialization there later.