kernel: Improve killbig() function
The mfreeseg() call in killbig() was wrong, as
mfreeseg() assumes to be called from a process
sharing the segment and relies on Segment.ref
to optimize the cross-processor TLB flushing.

One could try to fix that by incref()'ing the
segment before the mfreeseg() call, but we can
also just rely on the processes exiting itself
after getting killed.

Putting a comment above mfreeseg() to avoid this
mistake in the future.

For killing a process, refactor the code
from devproc.c into a killproc() function
in proc.c which handles cases for "Stopped" and
"Broken" processes on one place now.

As killbig() is called again (pages are not
freed immediately), we return when our victim
has: kp->procctl == Proc_exitbig. The caller
calls sched() in this case, giving these processes
a chance to exit themslfs.