kernel: improve imagecache locking
The assumptions are the following:

- attachimage() is called often (for each exec),
needing to acquire the imagealloc lock to walk the
list of active images.

- putimage() with a Image.ref dropping to 0 is rare,
only occuring when the system is under memory pressure
and we'r reclaiming pages from the image cache.

We used to acquire Image lock while holding
imagealloc lock as this was the only way to make sure
we do not race with putimage().

We want to avoid contending on the imagealloc lock here,
so that attachimage() can make progress and new processes
can be execed.

So the lock order is now changed, and we use Image.ref
to make sure the Image entry isnt freed after releasing
imagealloc lock.

This now means that putimage() must acquire imagealloc
lock and re-check that Image.ref == 0 before moving
the entry to the freelist. (holding allocimage lock
is the only way to make sure the Image isnt grabbed
out of the cache and used). And attachimage() never
holds imagealloc lock and Image lock at the same time.

In addition, simplify allocimage() call and move the
code attaching the first text segment to sysexec().
Recheck the active image list if we had to run
imagereclaim() to avoid duplicates.
Make attachimage() return a Image* with a unconditional
reference so the call is symmetic with putimage().