kernel: fix deadlock between sysexec() and killbig() (thanks ori)
During exec, we used to hold up->seglock while the stack
is being demand paged in.

In a low-memory situation, this leads to deadlock because
killbig() is calling procpagecount() for each process
to determine is physical memory usage. This function
acqires p->seglock. So killbig() can be stuck once it
hits a process that is starving on memory during faulting
in its stack.

This change makes sysexec release the seglock during its
its stack preparation phase, and re-aquire it once we
enter the "committ" phase (freeing our old stack).