kernel: attach text segment outside of up->seglock in sysexec()
The attachimage() call can be expensive as it might
need to reclaim some old images from the cache.

So do the attachimage() when up->seglock is released
(before the point of no return).

This means attachimage() can now trigger the pager
and any failures here just yield exec returning
with an error in the original process.