gefs: handle full disk and other errors more gracefully
if we had a full disk, or failed upserting, we could have
double-freed disk blocks on retry. This forces us to stop
writing, eliminating the chance for retry

This should also be safe after a restart -- because sync
refuses to write to disk after a failure, we will roll
back the free operations on failure, and start from the
last successful sync.

There may still be some issues if the second superblock
fails during a disk failure, but the window for issues
is much narrower.