gefs: tighten assertion on btupsert
When inserting a sequence of messages targeted at a key,
if the key doesn't already exist, we must start with an
Oinsert message to create the key. If a bug elsewhere in
the code leads to a key being updated without first being
created, we could end up with malformed keys being put
into the tree.

This shouldn't happen, but if it does, it's better to
crash here, before the data has been committed to disk,
than after we've already written it out and we noticed
that it was invalid.