kernel: cap the amount of work to reclaim pages form hash cain
When looking for free pages in a Image,
we want to skip over recently used pages
as they'r at the head of hash chains.

However, if the cache is very full, this
takes too long. So cap how deep we walk
into each chain and just take what we have.