stats: not on my bitsy
Commit 6a78bb2e4f00581d85de1c5f9949782de3dd8933 attempted to
address an issue with the temperature and battery percentage
calls being expensive through ACPI by creating a floor for how
often they are refreshed, however commit 55e5418e6945ec3275d34e04047ad706f6eb8e68
then changed the assumption for acpi to be mounted at /mnt/pm. While it may have
been planned to change the default mountpoint of of aux/acpi to /mnt/pm this has
yet to be done.

Attempting to guess at what the source the devices come from, and changing the global
floor on refreshing the values is a bit fragile and can have unintended side effects if
used with multiple machines. Instead let us just set reasonable values and document that
this is a potential problem with specifically ACPI. So the decision was made to make
battery refresh floor be 10 seconds, but cputemp to be kept at 1 second globally.
While it is unlikely that the battery percentage changes within 10 seconds, the
temperature very well could be changing on a second by second basis.

Additionally, there was a weird edge case in the code for specifically the bisty which
seemed to have a maximum charge percent of 184. Due to the how the code worked, with the
default aux/acpi invocation we would instead hit this bitsy case resulting in fully charged
thinkpads displaying as only ~1/2 charged. Let's just get rid of this mess.

Perhaps we should change the default mountpoint of aux/acpi to /mnt/pm, but for now we should
work well with the default invocation and document this quirk, while maintaining consistency for
wherever this may be mounted in the namespace.