libmemdraw: use fixed-point in memaffinewarp()
The previous version relied on libgeometry's invm(2) and xform(2) to do
its warping, which use double-precision floating-point numbers. They have
been imported and its operations replaced by fixed-point ones, so that it
can be included in the kernel for future use in draw(3) (most of the
architectures lack support for floating-point at the moment).

A new data structure, Warp, is now available to store the affine matrix;
along with a helper routine, mkwarp(), to translate libgeometry Matrices
into Warps that have the correct format necessary to do the transformations.