git/rebase: stop doing it.
git/rebase was a relatively clunky and error prone
script that would juggle around branches, make changes
to repository state, and ended up rather confusing.

it feels a lot better to use if we follow the example
of kill(1), where we print the commands that the user
may run, give them the opportunity to edit them, but
not take any action directly.

so let's do that.