libmemdraw: add image correlation and other improvements
i replaced the bit-shifting in some of the pixel sampling routines for
a single SWAR multiplication, which should be faster in most of the
arches we support; although i've only tested it on amd64 and arm64.

also wrote memimagecorrelate() which allows us to correlate and
convolve images with a filter or kernel.  it's limited to a single
filter for all of the color channels, since we don't have support for
images with more than 8bpc yet, and we use 32-bit fixed-point numbers
for every kernel component.

we still had a decision outstanding for memaffinewarp() as to how to
control the filter it uses for the transformation.  i wasn't getting
good results when trying to integrate convolution into the sampler, so
i added a parameter to choose whether you want a smooth, higher quality
result or a faster but blocky one.  a more knowledgeable user can then
do their own post-processing with the correlation operator.