libc: make needsrcquote the default
This removes footguns: The documentation said that quotes
were done 'in the style of rc', which was taken to mean that
they were quoted for rc.

Since it should always be safe to add more quotes when quoted
output is acceptable,  we should pick the option that the docs
imply and make it the default.